requirements.txt
nnmnkwii
numpy
pysptk
pyworld
pyyaml
tensorboardX
torch
scipy
